solution/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0
transformers>=4.48.0
huggingface_hub>=0.24.0
ultralytics>=8.2.0
opencv-python>=4.8.0
